requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
diff-match-patch==20241021
Django==5.2.12
django-import-export==4.4.0
et_xmlfile==2.0.0
idna==3.11
openpyxl==3.1.5
pillow==12.1.1
pyTelegramBotAPI==4.32.0
python-dotenv==1.2.2
qrcode==8.2
requests==2.32.5
sqlparse==0.5.5
tablib==3.9.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
